backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
flask-cors==4.0.1
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
chardet==5.2.0
bcrypt==4.1.2
PyJWT==2.8.0
flask-migrate==4.0.5
